Thu, Jul 6, 2023, 10:23 PM +06:00

+++ bluesheet/17.c
@@ -4,7 +4,7 @@
 
 int main(void)
 {
-    long arr[200000] = {1};
+    int arr[200000] = {1};
     int prime[200000] = {0};
     int p = 0;
